New Drive button and design task sidenav
1. I implemented new changes to the buttons on the milestones document.
I created a new button that allows for the user to bring files from
their Dropbox, which is another very popular drive.
2. I made new changes to the design task navbar.
    a. Added the sections: Project, Objective, Task. This should
clarify the navbar. One thing to think about moving forward is how to
include the radio buttons, checkboxes, etc. Perhaps they can be
versions of tasks?
    b. Added a feature where the user can click a button to have popup
where the user can read the description for a component.