Changed to multiprocessing.Lock()
Instead of threading.Lock(), TaskWorkerSerializer.create() now implements multiprocessing.Lock().
Additionally, the lock is implemented as a context manager to avoid unnecessary "try...finally" code. self.lock.release() will be called when the self.lock context manager is exited, whether through the return task_worker statement or due to an exception that causes the transaction to be rolled back entirely.